CCM-10055: Check if meshcli directory exists already

+++ scripts/get_mesh_cli.sh
@@ -1,35 +1,44 @@
 #!/bin/bash
 # Usage: ./import_package.sh <repo_url> <package_dir_in_repo> <target_dir_in_current_repo> [branch]
 
-REPO_URL="https://github.com/NHSDigital/comms-mgr.git"
-PACKAGE_DIR="packages/libs/mesh-cli"
-TARGET_DIR="/helpers/mesh-cli"
-BRANCH="${4:-main}"  # default branch
+#Check if directory exists already
+if [ -d "helpers/mesh-cli" ]; then
+  echo "Directory already exists"
+else
+    # Variables
+    REPO_URL="https://github.com/NHSDigital/comms-mgr.git"
+    PACKAGE_DIR="packages/libs/mesh-cli"
+    TARGET_DIR="/helpers/mesh-cli"
+    BRANCH="${4:-main}"
 
-TEMP_DIR=$(mktemp -d)
-cd "$TEMP_DIR" || exit 1
+    # Create a temporary directory
+    TEMP_DIR=$(mktemp -d)
+    cd "$TEMP_DIR" || exit 1
 
-# Initialize empty git repo quietly
-git init -q
+    # Initialize empty git repo quietly
+    git init -q
 
-# Add remote quietly
-git remote add origin "$REPO_URL"
+    # Add remote quietly
+    git remote add origin "$REPO_URL"
 
-# Enable sparse checkout quietly
-git sparse-checkout init --cone
-git sparse-checkout set "$PACKAGE_DIR" >/dev/null 2>&1
+    # Enable sparse checkout quietly
+    git sparse-checkout init --cone
+    git sparse-checkout set "$PACKAGE_DIR" >/dev/null 2>&1
 
-# Pull only the branch quietly
-git pull --quiet origin "$BRANCH"
+    # Pull only the branch quietly
+    git pull --quiet origin "$BRANCH"
 
-# Copy package into existing repo
-mkdir -p "$OLDPWD/$TARGET_DIR"
-cp -r "$PACKAGE_DIR/." "$OLDPWD/$TARGET_DIR/"
+    # Copy package into existing repo
+    mkdir -p "$OLDPWD/$TARGET_DIR"
+    cp -r "$PACKAGE_DIR/." "$OLDPWD/$TARGET_DIR/"
 
-# Cleanup
-cd "$OLDPWD"
-rm -rf "$TEMP_DIR"
+    # Cleanup
+    cd "$OLDPWD"
+    rm -rf "$TEMP_DIR"
 
-echo "$MESH_CLIENT_CONFIG" > helpers/mesh-cli/client_config.json
+    # Create a config file from environment variables
+    echo "$MESH_CLIENT_CONFIG" > helpers/mesh-cli/client_config.json
 
-pip install -e .
+    # Install the package
+    pip install -e .
+fi
